+++ CHANGES.md
@@ -1,3 +1,7 @@
+
+- `jsont.{bytesrw,brr}` export all requires.
+- `jsont` tool remove spurious dependency on `b0.std`.
+
 v0.1.0 2024-11-29 Zagreb
 ------------------------